Add option to notify only in background
This adds a -b/--only-in-background option that will only notify if the Jupyter notebook is in the background. This allows you to use something like `%autonotify -b -a 5`, which will notify on every cell execution over 5 seconds long, but only if you have switched windows or tabs. I think this is a pretty good experience as if the window is active you will already see that the cell has finished execution, and can avoid the extra notification.